LayoutTests/editing/pasteboard/copy-content-with-user-select-none-quirks-expected.txt

This tests copying excludes content with user-select: none.
To manually test, copy "hello world foo bar" below then paste.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS getSelection().toString().includes("hello") is true
PASS getSelection().toString().includes("world") is true
PASS getSelection().toString().includes("WebKit") is true
PASS getSelection().toString().includes("rocks") is true
PASS getSelection().toString().includes("because") is true
PASS getSelection().toString().includes("foo") is true
PASS getSelection().toString().includes("bar") is true
PASS event.clipboardData.getData("text/plain").includes("hello") is true
PASS event.clipboardData.getData("text/plain").includes("world") is true
PASS event.clipboardData.getData("text/plain").includes("WebKit") is true
PASS event.clipboardData.getData("text/plain").includes("rocks") is true
PASS event.clipboardData.getData("text/plain").includes("because") is true
PASS event.clipboardData.getData("text/plain").includes("foo") is true
PASS event.clipboardData.getData("text/plain").includes("bar") is true
PASS event.clipboardData.getData("text/html").includes("hello") is true
PASS event.clipboardData.getData("text/html").includes("world") is true
PASS event.clipboardData.getData("text/html").includes("<i") is true
PASS event.clipboardData.getData("text/html").includes("</i>") is true
PASS event.clipboardData.getData("text/html").includes("WebKit") is true
PASS event.clipboardData.getData("text/html").includes("<b ") is true
PASS event.clipboardData.getData("text/html").includes("</b>") is true
PASS event.clipboardData.getData("text/html").includes("rocks") is true
PASS event.clipboardData.getData("text/html").includes("<q>") is true
PASS event.clipboardData.getData("text/html").includes("</q>") is true
PASS event.clipboardData.getData("text/html").includes("because") is true
PASS event.clipboardData.getData("text/html").includes("<s>") is true
PASS event.clipboardData.getData("text/html").includes("</s>") is true
PASS event.clipboardData.getData("text/html").includes("<em>") is true
PASS event.clipboardData.getData("text/html").includes("</em>") is true
PASS event.clipboardData.getData("text/html").includes("foo") is true
PASS event.clipboardData.getData("text/html").includes("bar") is true
PASS successfullyParsed is true

TEST COMPLETE
hello world WebKit rocks because foo bar
hello world WebKit rocks because foo bar
